reafactor: @SpringBootApplication 와 @RestController 기능 분리
